requirements.txt
# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# Speech-to-text (local, fast)
faster-whisper>=1.0.0

# LLM - semantic slide detection
anthropic>=0.25.0

# PowerPoint control + content extraction
pywin32>=306
python-pptx>=0.6.23

# Prosodic analysis (Phase 3)
librosa>=0.10.0

# Utilities
pyyaml>=6.0
keyboard>=0.13.5

# Live display
rich>=13.0.0
plotext>=5.2.0
python-dotenv>=1.0.0